+++ src/main/java/ru/skypro/homework/model/entity/Comment.java
@@ -15,7 +15,7 @@ public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long pk;
+    private Integer pk;
 
     private LocalDateTime createdAt;
